remove agg from skia driver
Copy code from the agg driver so that the Skia
driver is no longer dependent on it.

The copied code doesn't in turn call
agg; it was in the agg driver for convenience.

Much of the copied code will eventually be replaced
by native Skia code, so there's not much point
in rewriting it to make it more efficient or
maintainable.

R=dsinclair@chromium.org,tsepez@chromium.org

Review URL: https://codereview.chromium.org/1865883002