Remove CPDF_Object::GetConstString and overrides
GetConstString() has sharp edges in that when applied to
a CPDF_Number, it must return null whereas GetString() returns
a the stringified number, because of the inability to
control the lifetime of the underlying allocated string.

Deleting this method showed several places where we
actually wanted a *String, not a *StringC, so we were
re-allocating a string we already had.

Review URL: https://codereview.chromium.org/1879683002