Only SolidBrush types are created, remove other brushes.
We only ever create a CFDE_SolidBrush, there is nothing which creates Texture,
Hatch or LinearGradient brushes. This Cl removes all the code that isn't
used for SolidBrush.

A followup Cl will rename SolidBrush to Brush and remove the Brush interface
in favour of the concrete class.

BUG=pdfium:468

Review URL: https://codereview.chromium.org/1877383002