Add dummy -j flags to pixel and javascript test runners
This is an intermediate step to creating a unified test
runner. We need to update Dr Memory to pass the -j 1 flag
to the pixel and javascript tests. In order to do so, we
need to have the flag available.

The flag will be enabled for pixel and javascript tests
when [1] lands.

1- https://codereview.chromium.org/1886753002/

Review URL: https://codereview.chromium.org/1882143005