Remove a few more char* members from structs/classes.
This is a result of looking for the anti-paterns /char\*\sm_/
and /m_.*=.*\.c_str/ which indicate that a class may be using
the contents of a string without extending the lifetime of
the underlying storage.

Along the way, change to uint8_t in fx_dib; this is unrelated but
avoids grep hits (it is binary, not chars anyways).

Also remove two string operators that make it easy to assign in
a manner that does not extend contents lifetime.

Review URL: https://codereview.chromium.org/1902953002