Upgrade build and clang.
The issue with //testing/multiprocess_func_list has been fixed in upstream
//build. This Cl upgrades //build to ToT.

In order to do so clang must also be updated. The Wno-undefined-var-template
is only passed if the pinned clang revision isn't 266460, which is the
currently pinned clang revision in chromium. This is there so that the bot
that builds with trunk clang gets this flag, but pinned clang doesn't.

We accidentally pickup the flag since our pinned clang is different from the
Chrome version. So, we must always roll //build and clang together when
changing deps or bad things may happen as there is an implicit dependency
due to compile flags.

BUG=pdfium:106

Review URL: https://codereview.chromium.org/1911483002