Allow overriding GN build flags
In order to enable XFA from GN builds we need to be able to override
pdf_enable_xfa. To do this a build_overrides/pdfium.gni file needs to be created
in Chrome. To use that, we have to create a build_overrides/pdfium.gni file
in PDFium that just uses the default values.

BUG=chromium:62400

Review-Url: https://codereview.chromium.org/1923333002