Relax a couple checks to allow certain non-standard PDF files.
Some non-standard PDF files misuse the size of cross reference table,
and reuse some object number which the old one is still in use. PDFium
can relax the reusing of xref objects only since it is not referred in
the pdf document. When the size of cross reference table is larger
than defined, PDFium will try to continue other than abort.

BUG=chromium:596947

Review-Url: https://codereview.chromium.org/1926823002