Reland of Allow overriding GN build flags
This reverts commit 038bf0b129e56212a6de0989dca1ae023bb17c5d.

In order to enable XFA from GN builds we need to be able to override
pdf_enable_xfa. To do this a build_overrides/pdfium.gni file needs to be created
in Chrome. To use that, //pdfium.gni reads the override values and sets them
into the defined arguments. We have to create a build_overrides/pdfium.gni file
in PDFium that sets the default overrides for PDFium builds.

BUG=chromium:62400

Review-Url: https://codereview.chromium.org/1936313002