Convert FWL_ERR into an enum class.
This Cl updates FWL_ERR to be an FWL_Error enum class. It removes FX_ERR and
replaces it with FWL_Error as well as the values were the same.

There were many places where we either returned other values for FWL_ERR
results.

This Cl is the same as: https://codereview.chromium.org/1943413002/ but I seem
to have messed up the base URL in that one.

TBR=tsepez@chromium.org

Review-Url: https://codereview.chromium.org/1952693003