Enable 'treating warnings as errors' for 64 bit Win builds
The reason to disable 'treating warnings as errors' earilier is due to
c4267 warnings, which we already disabled uniformly in pdfium.gyp.
Re-enable this now so that we can have the same level of compilation
warnings/errors on 32 bit and 64 bit builds.

Also fix two c4390 warnings on Windows release builds. Their warning
messages are "empty controlled statement found; is this the intent?"

Review-Url: https://codereview.chromium.org/1949303002