Use chromium_code standard for PDFium GYP compilation
Define and use chromium_code to be used in standalone PDFium GYP
build so that PDFium code can have more stringent warning level.

This is also enabled on GN build by default so that GYP and GN
builds can have consistent compilation results.

Also enable chromium_code for PDFium compilation in Chromium
since most of the warnings are cleared. The left ones are clearly
marked and will be addressed soon.

A few more clean-ups for the build:

-- Remove the suppression of sign-compare warnings for Clang since
the code is clean and the warning can be re-enabled.
-- Re-enable "treat warning as errors" on Mac
-- Add a flag to make GCC build works as well.

BUG=pdfium:29, pdfium:475

Review-Url: https://codereview.chromium.org/1985843002