XFA_WIDGETSTATUS cleanup.
This CL moves all of the XFA_WIDGETSTATUS values into a single enum instead
of multiple defines. The values are also normalized.

The name was updated to be XFA_WidgetStatus.

The XFA_WIDGETFILTER and XFA_LAYOUTSTATUS defines had to
match up to XFA_WIDGETSTATUS. This Cl replaces those so
we just have a single enum to work with.

Review-Url: https://codereview.chromium.org/1986503002