Cleanup XFA_HASHCODE usage.
This CL cleans up several uses of XFA_HASHCODE. The defines have been converted
into an enum.

For the SavePackage call the type is now used as the param instead of a string.
The callers pass in the correct type instead of doing an internal conversion.

The GetXFAObject accepting a string was removed as it was unused. The other
variant was changed to accept the XFA_HashCode type instead of uint32_t.

GetPackageData was removed as it is unused.

Review-Url: https://codereview.chromium.org/1989313002