Remove CFX_PrivateData from CPDF_Page
CFX_PrivateData served two purposes here:

The first was to hold an opaque pointer to the corresponding
page structure in the next higher layer, of which it knows
no details.  Introduce an empty CPDF_Page::View class to
represent this higher class, so as to get type safety while
preserving layering.

The second was to hold an opaque render context, which it
also happened to own.  Make this a CFX_Deletable to help
with management.

Also remove an unused inheritance from CFX_PrivateData
in CPDF_Annot.

Review-Url: https://codereview.chromium.org/2008553008