Remove CFX_PrivateData from CPDF_Document
Replace it with two generic slots for Links and Codec usage.
Since the codec is at a lower layer than the document, we
don't provide separate get/set methods, since having a
document upon which to call these would be a layering
violation.  Do the same for the Links for simplicity.

Review-Url: https://codereview.chromium.org/2005193003