Add opaque "layout item" to widgets for caller's use.
This avoids another use of CFX_PrivateData.  Note that in the old
code, we'd be calling through a m_pImpl onto the same underlying
object as we passed as the "key" argument when setting the value,
which explains why the get calls, happening one object lower, pass
the same argument as which they are being inovked against.

Review-Url: https://codereview.chromium.org/2010923002