Verify we have a CJBig2_Image before attempting use.
In CJBig2_SDDProc::decode_Arith we will set a SDNEWSYMS value to nullptr if the
height or width are 0. With the PDF from the bug, all of the decoders are set
to nullptr. Then, we call into CJBig2_TRDProc::decode_Arith and pull out
one of the nullptr decoders and attempt to use it, crashing.

This CL adds a check that we have a non-null decoder before attempting to use
the decoder.

BUG=pdfium:511

Review-Url: https://codereview.chromium.org/2048683002