fix skia + windows + gn
The Skia Windows build for PDFium differs from the
Skia Chromium build in that it uses FreeType within
PDFium and Direct Write within Chromium. This allows
Chrome to match the UI of Windows, and allows
PDFium to use FreeType to measure and draw.

When PDFium was updated to use gn, the settings
from Chrome were used as the basis for the PDFium
settings. Subsequently, PDFium built with Skia on
Windows drew text incorrectly as it used FreeType to
look up the font glyphs and Direct Write to draw them.

This fixes the gn files, and also fixes an error that
crept into the now less-used gyp files.

R=dsinclair@chromium.org
BUG=

Review-Url: https://codereview.chromium.org/2055353002