Speculative fix for FXJS_InitializeRuntime crash.
The call to GetGlobalObjectTemplate() in FXJS_InitializeRuntime() will call
into CFXJS_ObjDefinition::MaxID() which will call FXJS_PerIsolateData() on
the provided isolate. We, currently, create the isolate data after we make this
call which seems like it would lead to bad things. This CL moves the PerIsolate
data creation earlier in the process.

BUG=chromium:612918

Review-Url: https://codereview.chromium.org/2069763002