Fixup LoadImageInfo type checking.
The ::DetectImageType method does more then just detecting the image type, it
also sets up various needed structures to handle the decoding. Instead of
skipping the ::DetectImageType call this CL changes the code to return early if
the image check fails. This should allow us to stop working on images which do
not match the required data format.

BUG=chromium:621094

Review-Url: https://codereview.chromium.org/2085493002