Split the XFA_OBJECTTYPE enum into two parts.
Currently the object type and the node flags are both mixed into the single
XFA_OBJECTTYPE_* enum. These two things are un-related and should not share
a single type.  This Cl creates an XFA_ObjectType enum class and a
XFA_NodeFlag enum to hold the two types.

Accessors are added to determine if the flags are set (or called where they
already existed.

Review-Url: https://codereview.chromium.org/2083453003