Move CFX_ClipRgn and CFX_PathData into their own files.
This is the second CL in an attempt to split up the classes in fxge/include/fx_ge.h into their own files. CFX_ClipRgn is moved to core/fxge/ge because it is only used in core/fxge. The header for CFX_PathData is left in core/fxge/include since it is used elsewhere.

Review-Url: https://codereview.chromium.org/2216853004