Add support to Document::gotoNamedDest method.
Patch implements the Document's API gotoNamedDest, which is
part of the PDF specification [1], page 129, with the following
(short) description:

"Use this method to go to a named destination within the
PDF document".

[1] http://partners.adobe.com/public/developer/en/acrobat/sdk/5186AcroJS.pdf

"Named destination" is a common concept in the PDF world.
It can be used together with PDF's Links, Annotations, Bookmarks
and OpenActions, as well as an action per se, in case "this.gotoNamedDest"
is called directly.

Note that the implementation makes use of the existing hook
CPDFDoc_Environment::FFI_DoGoToAction, which ends up calling
out the embedder to actually handle it.
In case of Chromium, for instance, it calls PDFiumEngine::Form_DoGoToAction
which only handles for now the "page" property of the "destination".
Other properties, including zoom level, and scroll position
are ignored for the moment.

BUG=pdfium:492

Review-Url: https://codereview.chromium.org/2221823003