add native draw bitmap with alpha mask
PDFs draw bitmaps with alpha by supplying an
image alpha mask along side the image color data.

The alpha mask may not be the same scale as the
image data. Skia doesn't have a direct mapping
for this draw call, but can come close by using
the internal SkCompositeShader class.

This scales the alpha mask to match the size of
the image data, then puts both in the composite
shader to draw.

There are more things to do:

- Allow Alpha8 or Gray8 to draw scaled to
  a matching device (blit doesn't work today)

Temporary code marked with
a TODO implements this functionality.

- Refactor PDFium to allow SkComposeShader's
  components to take a matrix.

The latter goal will defer drawing so that PDFium
doesn't resize the mask, discarding resolution,
before Skia has a chance to record it.

Additionally fix drawing text with a negative
font size.

R=reed@google.com,dsinclair@chromium.org,herb@chromium.org

Review-Url: https://codereview.chromium.org/2182763002