Hidden annotations should not be drawn
Now that PDFium supports drawing of more annotation
types, it should also respect the "hidden" flag that
annotations might feature.
For instance, in IE/Acroread if an annotation is flagged as
"hidden" it does not get drawn.

CL adds a check for the specific "hidden" flag, not drawing
annotation that are flagged with it, in order to match IE + acrobat
reader behavior.

The "flags" definition can be seen by looking at "/F {value}"
syntax in a PDF file source, where {value} is an predefined
integer value.

Test: PDF files being added in [1].

[1] https://codereview.chromium.org/2239713003/

BUG=62625

Review-Url: https://codereview.chromium.org/2239853002