Use a map for app::m_Timers
Currently the timers is a vector. When we cancel a timer we have an O(n) operation
to remove the timer. If there are a lot of timers (which this test has > 16k) this can
take a long time. The impact is a lot lower in Release, but the test is very slow in Debug.

From Linux on waterfall:
  [ RUN      ] FPDFFormFillEmbeddertest.BUG_634716
  [       OK ] FPDFFormFillEmbeddertest.BUG_634716 (7855 ms)

From Linux try bot:
  [ RUN      ] FPDFFormFillEmbeddertest.BUG_634716
  [       OK ] FPDFFormFillEmbeddertest.BUG_634716 (431 ms)

From Linux XFA Rel GYP on waterfall:
  [ RUN      ] FPDFFormFillEmbeddertest.BUG_634716
  [       OK ] FPDFFormFillEmbeddertest.BUG_634716 (185 ms)

From Linux XFA Rel GYP try bot:
  [ RUN      ] FPDFFormFillEmbeddertest.BUG_634716
  [       OK ] FPDFFormFillEmbeddertest.BUG_634716 (72 ms)

Review-Url: https://codereview.chromium.org/2251333002