Add initial Document::getAnnot support
CL implements the first step in order to support
Annotations manipulation in PDFium: Document::getAnnot.

The method takes two arguments, an integer (page number)
and a string (annotation name).
When called, it iterates over the annotations on
the given page number, searching for the one whose name
matches the string in the second parameter.
If found, then an Annot instance (see Annot.cpp/g added by this
CL), is bound to a Javascript object and returned.

With the use cases described in bug [1] as an initial test case,
CL adds support to the following Annotation object properties:

- hidden
- name
- type

Idea is to keep evolving the implementation with more methods
and properties in follow up CLs.

[1] https://bugs.chromium.org/p/pdfium/issues/detail?id=492

BUG=pdfium:492

Review-Url: https://codereview.chromium.org/2260663002