Verify element exists before accessing.
Currently when the parser utility classes are outputting to a text buffer we do
not verify that an element from an array exists before accessing. We can have
null items in arrays (and dictionaries but the dictionary case is already
handled).

This Cl updates the code to check the element exists before attempting to use
the element.

BUG=chromium:641076

Review-Url: https://codereview.chromium.org/2292473004