Skip the channel if there is no data.
The JPX decoder needs to verify there is data associated with an image channel
before access. This was already done in one side of the if() but seems to be
missing from the other.

This Cl updates the loop to check the existance of channel data and to continue
iteration if none found.

BUG=chromium:637232

Review-Url: https://codereview.chromium.org/2291813002