Display content of the annotation when mouse hover.
Each annotation has its contents, and users should be able to see the
contents. In this patch, PDFium creates a Popup annotation for each
annotation and stores the author and the content. When a user mouse
hover on the annotation, PDFium draws the corresponding Popup annotation
and displays the content.

Also, roll DEPS for testing/corpus to 5867fa6.

BUG=62625

Review-Url: https://codereview.chromium.org/2273893002