Fix leaks due to created popup annotations
When we create popup annotations, we also create the dictionary
associated with it. For regular annotations, the dictionary
associated with an annotation is not owned by annotation,
and will be released separately. But our created dictionary is not
associated with any other data structure, it would be leaked if not
released by the associated annotation.

Add a boolean to indicate the ownership to the dictionary, and release
the owned dictionary during the destruction of an annotation.

BUG=pdfium:242

Review-Url: https://codereview.chromium.org/2301613002