Do not re-create the page view when accessing from Document::removeField.
When removing a field from the document we can force the recreation of a
CPDFSDK_PageView by passing |true| as the last parameter to |GetPageView|. This
will force a new page view to be created so we can delete a field from it and,
we then end up with a dangling PageView.

This CL passes |false| to |GetPageView| and skips removing the widget if we
do not receive a pageView back from the document.

BUG=chromium:644219

Review-Url: https://codereview.chromium.org/2312023004