Pass CFX_WideString further down widget callers
Avoid a couple of places where we copy the raw string back into a
brand-new widestring.

There are a few places where the difference between a null
ptr and an empty string control the logic, and I left these
as-is.  Other places can just take the string by const ref.

Review-Url: https://codereview.chromium.org/2323493002