Reland of Fix leaked internal font
Reland of Fix leaked internal font (patchset #2 id:60001 of
https://codereview.chromium.org/2297303004/ )

In CFGAS_FontMgrImp::LoadFont(), a new internal font is created which
is never released. It needs to be correctly marked as internal font to
be released. Fix this by adding a new method to take the ownership
of the font and mark it as internal font properly.

The previous revert was caused by memory management errors
which were fixed at https://codereview.chromium.org/2322043002/

BUG=pdfium:242

Review-Url: https://codereview.chromium.org/2320213002